+++ src/extensions/server.rs
@@ -1,4 +1,10 @@
+use std::{future::Future, net::SocketAddr};
+
 use async_trait::async_trait;
+use jsonrpsee::server::{
+    middleware::proxy_get_request::ProxyGetRequestLayer,
+    {RandomStringIdProvider, RpcModule, ServerBuilder, ServerHandle},
+};
 use serde::Deserialize;
 
 use crate::{extension::Extension, utils::TypeRegistryRef};
@@ -37,5 +43,30 @@ impl Extension for Server {
 impl Server {
     pub fn new(config: ServerConfig) -> Self {
         Self { config }
+    }
+
+    pub async fn create_server<Fut: Future<Output = anyhow::Result<(RpcModule<()>)>>>(
+        &self,
+        builder: impl FnOnce() -> Fut,
+    ) -> anyhow::Result<(SocketAddr, ServerHandle)> {
+        let service_builder =
+            tower::ServiceBuilder::new().option_layer(self.config.health.as_ref().map(|h| {
+                ProxyGetRequestLayer::new(h.path.clone(), h.method.clone())
+                    .expect("Invalid health config")
+            }));
+
+        let server = ServerBuilder::default()
+            .set_middleware(service_builder)
+            .max_connections(self.config.max_connections)
+            .set_id_provider(RandomStringIdProvider::new(16))
+            .build((self.config.listen_address.as_str(), self.config.port))
+            .await?;
+
+        let module = builder().await?;
+
+        let addr = server.local_addr()?;
+        let server = server.start(module)?;
+
+        Ok((addr, server))
     }
 }
