+++ src/lib.rs
@@ -1,10 +1,10 @@
 pub mod config;
+pub mod extension;
+pub mod extensions;
 pub mod logger;
 pub mod middleware;
 pub mod middlewares;
-// pub mod server;
-pub mod extension;
-pub mod extensions;
+pub mod server;
 pub mod utils;
 
 #[cfg(test)]
